+++ database/migrations/2022_02_05_152847_remove_old_wake_dependencies.php
@@ -1,0 +1,28 @@
+<?php
+
+use App\Services\DependencyService;
+use Illuminate\Database\Migrations\Migration;
+
+class RemoveOldWakeDependencies extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        DependencyService::deleteDependency('DEPENDENCY_WAKE');
+        DependencyService::deleteDependency('DEPENDENCY_RECAT');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        //
+    }
+}